Bug fix and reorder gpio pins.
Changed the way RGB values are split up. This method is probably not the best way to accomplish this,
but it works and I will still be able to make sence of it if I look at it a year from now.

The order of the pins now puts them in a much more logical order, with ground followed by red, green, and
blue in that order. The pervious numerical ordering put the red in the middle.