Fixed bugs in OTA routines.
Moved the call to ArduinoOTA.handle(); because the server.handleClient() was creating a race
condition that did not allow the OTA to sucessfully complete an update.

Fixed a bug in the OTA visual feedback that was not properly indicating upload progress. Now
during an update the moon starts at full blue and fades out as progress continues and turns green
on completion or red on error.