Fixed bug in first turning on the LED.
There was a bug where the color of the led was not set properly until after a client connected
to the network (the second time color is set. at the end of the setup, which shopuld probably be
removed now that it is done right the first time...) By putting it a little later in the setup
it gives the desired results. It was most likely setting the color too soon after setting up the
output, and not giving time for the gpio to be initalized in the hardware.