All systems go!
This is a fairly major rework. The save function finally works properly. Websockets now
communicate in both directions. This allows the client to always display the correct color
for the moon, even in rainbow mode! Some unnecessary image files were removed to save space.
CSS styles have been adusted and some bugs were removed there.

There does seem to be an ocassional bug causing the websocket to be prematurely closed but it
is not being reported by either the client or the server so I am not sure which is dropping
the ball there.