Serial.print lines commented out.
Due to the current size of the compiled binary Serial.print()
lines have been commented out. In production they are not useful
anyway because the ESP-01 does not have enough available pins to
connect a serial output anyway. If any are needed for debugging
they can be uncommented, but be careful, making the program size
much bigger might break OTA updates.