Fixed bug theat led to abandoning websockets instead of closing them.
Webockets are now closed before leaving the main page.  This will improve
stability, but also means that any addition pagesal that need the use of
websockets added in the future should be loaded as a frame (?) in the main
window.
Added a batery check function that will turn the moon red and fade in and
out to indicate a recharge is necessary when the battery in down to the
point that color levels become unstable.