Removed blocking wait in WiFi setup routine.
Removed the section of startWiFi() that waited for a connection
before finishing the setup and proceeding to the main loop.
Because the current implementation is only using AP mode there
is no need to wait for a connection, and this also prevented
running in Rainbow mode as the default because the loop - where
the color rotation needs to happen was not reached in AP mode
until after a client connected to the wifi network.
Logging to the javascipt web interface has also been reduced.
The debug logging lines are still there, only commented out
just in case more logging is needed for debugging.