Improved preference storage backend
This changes the preference storage backend, adding a pair of bits to each
saved EEPROM preference entry to check if the slot hold a valid entry.
Storage structures are in place for storing color settings and mode, as well
as network configuration info. There is one slot for AP moded `ssid` and `psk`
as well as slots to store up to three STA mode network `ssid` and `psk` pairs.
Mechanisms are alredy in place to load these configurations into the WiFiMulti
client, but the configuration and storage frontend for the UI still needs to be
added.

Signed-off-by: Winford (Uncle Grumpy) <winford@object.stream>