Fix bug in javascript
TYhis fixes a bug I accidentally introduced, when trying to make "standard",
the JavaScript linter happy. It helped my find bugs, but is not made with
vanilla javascript in mind, as with **all** other javascript linters I have
encountered it assumes you are using `node`, not JS in an HTML webpage.

I would love a vanilla JavaScript linter, but the only I have found yet are
ancient perl scripts.

Signed-off-by: Winford (Uncle Grumpy) <winford@object.stream>